+++ resources/js/pages/auth/SetPassword.vue
@@ -3,13 +3,13 @@
   import {t} from '@craftcms/cp';
   import {usePage} from '@inertiajs/vue3';
   import AuthBase from '@/common/layouts/AuthBase.vue';
-  import {store} from '@actions/Auth/SetPasswordController';
   import '@/modules/auth/components/set-password/set-password-form.js';
 
   const props = defineProps<{
     uid: string;
     code: string;
     newUser: boolean;
+    action: string;
   }>();
 
   const page = usePage<{
@@ -18,10 +18,6 @@
 
   const title = computed(() =>
     props.newUser ? t('Set Your Password') : t('Set Your New Password')
-  );
-
-  const action = computed(
-    () => store(undefined, {query: {uid: props.uid, code: props.code}}).url
   );
 </script>